+++ mysensors/mysensors.py
@@ -27,7 +27,7 @@ class Gateway(object):
         if persistence:
             self._load_sensors()
         if protocol_version == "1.4":
-            _const = __import__("mysensors.const", globals(), locals(), ['Internal', 'MessageType'], 0)
+            _const = __import__("mysensors.const_14", globals(), locals(), ['Internal', 'MessageType'], 0)
         global Internal, MessageType
         Internal = _const.Internal
         MessageType = _const.MessageType
